moving into cycloid-resource with feature field
Now this resource have a "feature" configuration field.
This field is here to define the behavior of the resource.

The previous implementation had few issues with Concourse logic when used
for infrapolicy and terracost.

Now both are concidered as 2 features and have their own version and display system.

Note the user will now have to define 2 resource of this type to be able to use both Cycloid
features in the pipeline